Move geospatial tests to plugin module
