+++ core/trino-main/src/test/java/io/trino/plugin/geospatial/TestPruneSpatialJoinColumns.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.sql.planner.iterative.rule;
+package io.trino.plugin.geospatial;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -20,6 +20,7 @@ import io.trino.sql.ir.FunctionCall;
 import io.trino.sql.ir.SymbolReference;
 import io.trino.sql.planner.Symbol;
 import io.trino.sql.planner.assertions.PlanMatchPattern;
+import io.trino.sql.planner.iterative.rule.PruneSpatialJoinColumns;
 import io.trino.sql.planner.iterative.rule.test.BaseRuleTest;
 import io.trino.sql.planner.plan.Assignments;
 import io.trino.sql.planner.plan.SpatialJoinNode;
@@ -28,11 +29,11 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Optional;
 
+import static io.trino.plugin.geospatial.TestPruneSpatialJoinChildrenColumns.TEST_ST_DISTANCE_FUNCTION;
 import static io.trino.sql.ir.ComparisonExpression.Operator.LESS_THAN_OR_EQUAL;
 import static io.trino.sql.planner.assertions.PlanMatchPattern.spatialJoin;
 import static io.trino.sql.planner.assertions.PlanMatchPattern.strictProject;
 import static io.trino.sql.planner.assertions.PlanMatchPattern.values;
-import static io.trino.sql.planner.iterative.rule.TestPruneSpatialJoinChildrenColumns.TEST_ST_DISTANCE_FUNCTION;
 
 public class TestPruneSpatialJoinColumns
         extends BaseRuleTest
